Update rename + fixup imports
